Created component for Dine In or Delivery forms, removed them from Orders component
